Refactoring to strategy per source
